Implemented find userById middleware
