+++ app.js
@@ -7,6 +7,7 @@ const expressValidator = require("express-validator");
 
 require("dotenv").config();
 // import routes
+const authRoutes = require("./routes/auth");
 const userRoutes = require("./routes/user");
 
 // app
@@ -26,6 +27,7 @@ app.use(cookieParser());
 app.use(expressValidator());
 
 // routes middlewares
+app.use("/api", authRoutes);
 app.use("/api", userRoutes);
 
 const port = process.env.PORT || 8000;
